im tapping out
The spawntimer script was useless and I deleted it. right now the issue is the the enemyspawner script and how it interacts with the object pool. It currently will instantiate an enemy or two, but then it wont turn them on, and will only "spawn" a few enemies.

I added a debug to spawn enemies with spacebar.